+++ CHANGELOG.md
@@ -59,6 +59,10 @@ Enhancements
 * `StaticVec` is changed to keep three items inline instead of four.
 
 
+Version 1.0.7
+=============
+
+
 Version 1.0.6
 =============
 
@@ -66,6 +70,7 @@ Bug fixes
 ---------
 
 * Eliminate unnecessary property write-back when accessed via a getter since property getters are assumed to be _pure_.
+* Writing to a property of an indexed valued obtained via an indexer now works properly by writing back the changed value via an index setter.
 
 Enhancements
 ------------